Harden twitter digest collectors
