+++ twitter-digest/scripts/run_daily_digest.py
@@ -4,7 +4,6 @@
 from __future__ import annotations
 
 import argparse
-import base64
 import json
 import os
 import platform
@@ -12,19 +11,16 @@ import subprocess
 import shlex
 import sys
 import tempfile
-import time
-import urllib.parse
-import urllib.request
 from pathlib import Path
 
+from api_config_store import API_CONFIG_PATH, load_api_config, refresh_oauth_token_if_needed
 from digest_context import build_current_context_from_file
+from script_utils import display_path
 
 
 STATE_DIR = Path(__file__).resolve().parents[1] / ".state"
 CONFIG_PATH = STATE_DIR / "config.json"
-API_CONFIG_PATH = STATE_DIR / "api_config.json"
 DEFAULT_OUT_DIR = STATE_DIR / "run"
-TOKEN_URL = "https://api.x.com/2/oauth2/token"
 
 
 def parse_args() -> argparse.Namespace:
@@ -66,81 +62,6 @@ def load_config() -> dict:
         return {}
 
 
-def load_api_config() -> dict:
-    if not API_CONFIG_PATH.exists():
-        return {}
-    try:
-        data = json.loads(API_CONFIG_PATH.read_text(encoding="utf-8"))
-    except Exception:
-        return {}
-    if not isinstance(data, dict):
-        return {}
-    user_id = str(data.get("user_id") or "")
-    if user_id and not user_id.isdigit() and not data.get("handle"):
-        data["handle"] = user_id.lstrip("@")
-        data["user_id"] = ""
-    return data
-
-
-def save_api_config(config: dict) -> None:
-    API_CONFIG_PATH.parent.mkdir(parents=True, exist_ok=True)
-    try:
-        API_CONFIG_PATH.parent.chmod(0o700)
-    except PermissionError:
-        pass
-    API_CONFIG_PATH.write_text(json.dumps(config, ensure_ascii=False, indent=2) + "\n", encoding="utf-8")
-    try:
-        API_CONFIG_PATH.chmod(0o600)
-    except PermissionError:
-        pass
-
-
-def refresh_oauth_token_if_needed(config: dict) -> dict:
-    refresh_token = str(config.get("refresh_token") or "")
-    client_id = str(config.get("client_id") or "")
-    if not refresh_token or not client_id:
-        return config
-    try:
-        expires_at = int(str(config.get("expires_at") or "0"))
-    except ValueError:
-        expires_at = 0
-    if expires_at and expires_at - int(time.time()) > 300:
-        return config
-    form = {
-        "grant_type": "refresh_token",
-        "refresh_token": refresh_token,
-        "client_id": client_id,
-    }
-    body = urllib.parse.urlencode(form).encode("utf-8")
-    headers = {"Content-Type": "application/x-www-form-urlencoded"}
-    client_secret = str(config.get("client_secret") or "")
-    if client_secret:
-        encoded = base64.b64encode(f"{client_id}:{client_secret}".encode("utf-8")).decode("ascii")
-        headers["Authorization"] = f"Basic {encoded}"
-    request = urllib.request.Request(TOKEN_URL, data=body, headers=headers, method="POST")
-    try:
-        with urllib.request.urlopen(request, timeout=30) as response:
-            token = json.loads(response.read().decode("utf-8"))
-    except Exception as exc:
-        print(f"Saved OAuth token could not be refreshed: {exc}. Falling back to current token.", flush=True)
-        return config
-    access_token = str(token.get("access_token") or "")
-    if not access_token:
-        return config
-    now = int(time.time())
-    updated = dict(config)
-    updated["bearer_token"] = access_token
-    if token.get("refresh_token"):
-        updated["refresh_token"] = str(token.get("refresh_token"))
-    updated["token_type"] = str(token.get("token_type") or updated.get("token_type") or "bearer")
-    updated["expires_in"] = str(token.get("expires_in") or "")
-    updated["issued_at"] = str(now)
-    updated["expires_at"] = str(now + int(token.get("expires_in") or 0)) if token.get("expires_in") else ""
-    save_api_config(updated)
-    print("Refreshed saved X OAuth access token.", flush=True)
-    return updated
-
-
 def save_config(handle: str | None, account_name: str | None) -> None:
     CONFIG_PATH.parent.mkdir(parents=True, exist_ok=True)
     config = load_config()
@@ -149,13 +70,6 @@ def save_config(handle: str | None, account_name: str | None) -> None:
     if account_name:
         config["account_name"] = account_name
     CONFIG_PATH.write_text(json.dumps(config, ensure_ascii=False, indent=2) + "\n", encoding="utf-8")
-
-
-def display_path(path: Path) -> str:
-    try:
-        return "~/" + str(path.expanduser().resolve().relative_to(Path.home()))
-    except ValueError:
-        return str(path)
 
 
 def open_config_in_terminal(extra_args: list[str]) -> bool:
