+++ twitter-digest/scripts/script_utils.py
@@ -1,0 +1,20 @@
+"""Small shared utilities for twitter-digest scripts."""
+
+from __future__ import annotations
+
+from pathlib import Path
+
+
+def display_path(path: Path) -> str:
+    try:
+        return "~/" + str(path.expanduser().resolve().relative_to(Path.home()))
+    except ValueError:
+        return str(path)
+
+
+def ensure_private_dir(path: Path) -> None:
+    path.mkdir(parents=True, exist_ok=True)
+    try:
+        path.chmod(0o700)
+    except PermissionError:
+        pass
